Remove useless AbstractXXXRuleConfigurationChecker and move logic to implementation class
