+++ features/encrypt/core/src/main/java/org/apache/shardingsphere/encrypt/checker/EncryptRuleConfigurationChecker.java
@@ -17,25 +17,70 @@
 
 package org.apache.shardingsphere.encrypt.checker;
 
+import com.google.common.base.Preconditions;
+import com.google.common.base.Strings;
 import org.apache.shardingsphere.encrypt.api.config.EncryptRuleConfiguration;
+import org.apache.shardingsphere.encrypt.api.config.rule.EncryptColumnRuleConfiguration;
 import org.apache.shardingsphere.encrypt.api.config.rule.EncryptTableRuleConfiguration;
 import org.apache.shardingsphere.encrypt.constant.EncryptOrder;
+import org.apache.shardingsphere.infra.config.rule.checker.RuleConfigurationChecker;
+import org.apache.shardingsphere.infra.rule.ShardingSphereRule;
 
+import javax.sql.DataSource;
 import java.util.Collection;
+import java.util.Map;
 
 /**
  * Encrypt rule configuration checker.
  */
-public final class EncryptRuleConfigurationChecker extends AbstractEncryptRuleConfigurationChecker<EncryptRuleConfiguration> {
+public final class EncryptRuleConfigurationChecker implements RuleConfigurationChecker<EncryptRuleConfiguration> {
     
     @Override
-    protected Collection<String> getEncryptors(final EncryptRuleConfiguration config) {
-        return config.getEncryptors().keySet();
+    public void check(final String databaseName, final EncryptRuleConfiguration config, final Map<String, DataSource> dataSourceMap, final Collection<ShardingSphereRule> rules) {
+        checkTableConfiguration(databaseName, config.getTables(), config.getEncryptors().keySet());
     }
     
-    @Override
-    protected Collection<EncryptTableRuleConfiguration> getTables(final EncryptRuleConfiguration config) {
-        return config.getTables();
+    private void checkTableConfiguration(final String databaseName, final Collection<EncryptTableRuleConfiguration> tables, final Collection<String> encryptors) {
+        for (EncryptTableRuleConfiguration each : tables) {
+            for (EncryptColumnRuleConfiguration column : each.getColumns()) {
+                checkCipherColumnConfiguration(databaseName, encryptors, column);
+                checkAssistColumnConfiguration(databaseName, encryptors, column);
+                checkLikeColumnConfiguration(databaseName, encryptors, column);
+            }
+        }
+    }
+    
+    private void checkCipherColumnConfiguration(final String databaseName, final Collection<String> encryptors, final EncryptColumnRuleConfiguration column) {
+        Preconditions.checkState(!Strings.isNullOrEmpty(column.getCipherColumn()),
+                "Cipher column of `%s` can not be null in database `%s`.", column.getLogicColumn(), databaseName);
+        Preconditions.checkState(!Strings.isNullOrEmpty(column.getEncryptorName()),
+                "Encryptor name of `%s` can not be null in database `%s`.", column.getLogicColumn(), databaseName);
+        Preconditions.checkState(encryptors.contains(column.getEncryptorName()),
+                "Can not find encryptor `%s` in database `%s`.", column.getEncryptorName(), databaseName);
+    }
+    
+    private void checkAssistColumnConfiguration(final String databaseName, final Collection<String> encryptors, final EncryptColumnRuleConfiguration column) {
+        if (Strings.isNullOrEmpty(column.getAssistedQueryColumn()) && Strings.isNullOrEmpty(column.getAssistedQueryEncryptorName())) {
+            return;
+        }
+        Preconditions.checkState(!Strings.isNullOrEmpty(column.getAssistedQueryColumn()),
+                "Assisted query column of `%s` can not be null in database `%s`.", column.getLogicColumn(), databaseName);
+        Preconditions.checkState(!Strings.isNullOrEmpty(column.getAssistedQueryEncryptorName()),
+                "Assisted query encryptor name of `%s` can not be null in database `%s`.", column.getLogicColumn(), databaseName);
+        Preconditions.checkState(encryptors.contains(column.getAssistedQueryEncryptorName()),
+                "Can not find assisted query encryptor `%s` in database `%s`.", column.getAssistedQueryEncryptorName(), databaseName);
+    }
+    
+    private void checkLikeColumnConfiguration(final String databaseName, final Collection<String> encryptors, final EncryptColumnRuleConfiguration column) {
+        if (Strings.isNullOrEmpty(column.getLikeQueryColumn()) && Strings.isNullOrEmpty(column.getLikeQueryEncryptorName())) {
+            return;
+        }
+        Preconditions.checkState(!Strings.isNullOrEmpty(column.getLikeQueryColumn()),
+                "Like query column of `%s` can not be null in database `%s`.", column.getLogicColumn(), databaseName);
+        Preconditions.checkState(!Strings.isNullOrEmpty(column.getLikeQueryEncryptorName()),
+                "Like query encryptor name of `%s` can not be null in database `%s`.", column.getLogicColumn(), databaseName);
+        Preconditions.checkState(encryptors.contains(column.getLikeQueryEncryptorName()),
+                "Can not find like query encryptor `%s` in database `%s`.", column.getLikeQueryEncryptorName(), databaseName);
     }
     
     @Override
